+++ example/lib/main.dart
@@ -14,7 +14,15 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  String _platformVersion = 'Unknown';
+  String _tokboxSessionId = 'YOUR_SESSION_ID';
+  String _tokboxToken = 'YOUR_TOKEN';
+  String _tokboxApiKey = 'YOUR_API_KEY';
+  String _publishId = 'PUBLISH_ID';
+
+  bool _sessionInited = false;
+  bool _isPublishing = false;
+
+  int _pluginViewId = -1;
 
   @override
   void initState() {
@@ -22,35 +30,110 @@ class _MyAppState extends State<MyApp> {
     initPlatformState();
   }
 
-  // Platform messages are asynchronous, so we initialize in an async method.
-  Future<void> initPlatformState() async {
-    String platformVersion;
-    // Platform messages may fail, so we use a try/catch PlatformException.
-    try {
-      platformVersion = await FlutterVonageVideo.platformVersion;
-    } on PlatformException {
-      platformVersion = 'Failed to get platform version.';
-    }
-
-    // If the widget was removed from the tree while the asynchronous platform
-    // message was in flight, we want to discard the reply rather than calling
-    // setState to update our non-existent appearance.
-    if (!mounted) return;
-
+  Future<void> _initSession() async {
+    String ret = await FlutterVonageVideo.initSession(_tokboxSessionId, _tokboxToken, _tokboxApiKey);
     setState(() {
-      _platformVersion = platformVersion;
+      _sessionInited = true;
+      _isPublishing = false;
     });
+  }
+
+  Future<void> _publishStream() async {
+    String ret = await FlutterVonageVideo.publishStream(_publishId, _pluginViewId);
+    setState(() {
+      _isPublishing = true;
+    });
+  }
+
+  Future<void> _unpublishStream() async {
+    String ret = await FlutterVonageVideo.unpublishStream();
+    setState(() {
+      _isPublishing = false;
+    });
+  }
+
+  Widget _buildPublisher(var context) {
+    String viewType = 'flutter-vonage-video-publisher';
+    Map<String, dynamic> creationParams = <String, dynamic> {};
+    if (defaultTargetPlatform == TargetPlatform.android) {
+      return PlatformViewLink(
+        viewType: viewType,
+        surfaceFactory: (BuildContext context, PlatformViewController controller) {
+          return AndroidViewSurface(
+            controller: controller,
+            gestureRecognizers: const <Factory<OneSequenceGestureRecognizer>>{},
+            hitTestBehavior: PlatformViewHitTestBehavior.opaque,
+          );
+        },
+        onCreatePlatformView: (PlatformViewCreationParams params) {
+          return PlatformViewsService.initSurfaceAndroidView(
+            id: params.id,
+            viewType: viewType,
+            layoutDirection: TextDirection.ltr,
+            creationParams: creationParams,
+            creationParamsCodec: StandardMessageCodec(),
+          )
+          ..addOnPlatformViewCreatedListener(params.onPlatformViewCreated)
+          ..create();
+        },
+      );
+    } else if(defaultTargetPlatform == TargetPlatform.iOS) {
+      return UiKitView(
+        viewType: viewType,
+        layoutDirection: TextDirection.ltr,
+        creationParams: creationParams,
+        creationParamsCodec: const StandardMessageCodec(),
+        onPlatformViewCreated: (int id) {
+          _pluginViewId = id;
+        },
+      );
+    }
+  }
+
+  @override
+  void dispose() {
+    if (_isPublishing) {
+      _unpublishStream();
+    }
+    if (_sessionInited) {
+      FlutterVonageVideo.endSession();
+    }
+    super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
+    Widget _buttonPublish = SizedBox.shrink();
+    if (_sessionInited && !_isPublishing) {
+      _buttonPublish = ElevatedButton(
+        child: Text('Publish Stream'),
+        onPressed: _publishStream,
+      );
+    }
+
+    Widget _buttonUnpublish = SizedBox.shrink();
+    if (_sessionInited && _isPublishing) {
+      _buttonUnpublish = ElevatedButton(
+        child: Text('Unpublish Stream'),
+        onPressed: _unpublishStream,
+      );
+    }
+
     return MaterialApp(
       home: Scaffold(
         appBar: AppBar(
-          title: const Text('Plugin example app'),
+          title: const Text('Vonage Video example app'),
         ),
-        body: Center(
-          child: Text('Running on: $_platformVersion\n'),
+        body: ListView(
+          children: <Widget> [
+            _buttonPublish,
+            _buttonUnpublish,
+            Container(
+              width: 600,
+              height: 600,
+              child: _buildPublisher(context),
+            ),
+          ]
         ),
       ),
     );